Separate logic functions from the ztore
