+++ src/todo/Todo.jsx
@@ -1,10 +1,10 @@
 import { ztore } from '../zustand/ztore.js'
 
 export const Todo = () => {
-  const { alarmInterval } = ztore()
+  const { currentAlarmInterval } = ztore()
   return (
     <div className="todo">
-      TODO alarm interval : {alarmInterval.current}
+      TODO alarm interval : {currentAlarmInterval}
     </div>
   )
 }
